requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
docker-pycreds==0.4.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///C:/ci/gmpy2_1645455782955/work
grpcio==1.62.2
huggingface-hub==0.22.2
hydra-core==1.2.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib_metadata==7.1.0
intel-openmp==2021.4.0
Jinja2 @ file:///C:/b/abs_f7x5a8op2h/croot/jinja2_1706733672594/work
joblib==1.4.0
lightly==1.2.22
lightly-utils==0.0.2
lightning-bolts==0.5.0
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.5
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numpy @ file:///C:/b/abs_54abayvc9j/croot/numpy_and_numpy_base_1682520598361/work
omegaconf==2.3.0
packaging==24.0
pandas==2.2.2
pathtools==0.1.2
pillow==10.3.0
plotly==5.11.0
promise==2.3
protobuf==3.20.1
psutil==5.9.8
pyDeprecate==0.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.6.4
pytz==2024.1
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
safetensors==0.4.3
scikit-learn==1.2.0
scipy==1.13.0
sentry-sdk==2.0.1
setproctitle==1.3.3
shortuuid==1.0.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tbb==2021.12.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.3.0
threadpoolctl==3.4.0
timm==0.9.16
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.3.2
torchvision==0.18.0
tqdm==4.66.2
typing_extensions @ file:///C:/b/abs_72cdotwc_6/croot/typing_extensions_1705599364138/work
tzdata==2024.1
urllib3 @ file:///C:/b/abs_a9n398cvep/croot/urllib3_1700840517182/work
wandb==0.12.11
Werkzeug==3.0.2
yarl==1.9.4
yaspin==3.0.2
zipp==3.18.1
